requirements.txt
configparser==3.7.4
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.0.2
django-debug-toolbar==1.11
django-rest-passwordreset==1.0.0
djangorestframework==3.9.2
pytz==2019.1
sqlparse==0.3.0
whitenoise==3.2.1
